+++ Changes.md
@@ -3,6 +3,7 @@ Version Information
 Version 3.8.1.2 - TBR
 ---------------------
 1. Fix header opacity settings not resetting the theme cache.
+2. Added swatch custom colour detection.
 
 Version 3.8.1.1 - 04/03/21
 --------------------------